Small changes for equals/hashcode and a few documentation fixes

+++ examples/org/dyn4j/examples/ExampleGraphics2D.java
@@ -373,7 +373,7 @@ public class ExampleGraphics2D extends JFrame {
         // set the last time
         this.last = time;
     	// convert from nanoseconds to seconds
-    	double elapsedTime = (double)diff / NANO_TO_BASE;
+    	double elapsedTime = diff / NANO_TO_BASE;
         // update the world with the elapsed time
         this.world.update(elapsedTime);
 	}
